actions: Consolidate separate test actions into one
Previously we split up our testing CI into separate GitHub
Actions that run in parallel. But because each action has
to perform the same installation and build steps, I'm not
sure if we are actually saving any time.

This is an attempt to speed things by having all of our
Actions share the installation and build steps.