Makefile: Define recursive directory dependencies with a Python script
Previously, we were querying all files in a given directory using
`git ls-files`. This was an easy, cross-platform way to do it
because we could assume that the user always had Git installed--whether
they were on Windows, macOS, or Linux.

But it was hard to get `git ls-files` to exactly do what we wanted,
so we are going to use a simple cross-platform Python script
to list all of the files and directories under a given path.